+++ index.tsx
@@ -1,5 +1,4 @@
-import CONSTANT from './src/components/utils/constant'
 import functions from './src/components/utils/functions'
 import UI from './src/components/controls/index'
 
-export { UI, functions, CONSTANT }
+export { UI, functions }